requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.1.0
python-dotenv==0.20.0
requests==2.28.1

